Move overlay from template to component folder
