--- services/frontend/src/components/Menu/menu.css
@@ -1,29 +1,0 @@
-.menu {
-	width: 100%;
-	height: 100%;
-	flex-direction: row;
-	gap: 32px;
-	border-radius: 0;
-}
-
-.menu a {
-	text-decoration: none;
-}
-
-.menu .pop-hover-container {
-	gap: 8px;
-}
-
-
-.menu p {
-	opacity: 1;
-	transition: opacity 0.3s;
-}
-
-.menu .button.active {
-	color: var(--primary-color);
-}
-
-.menu-pophover {
-	width: max-content
-}
