+++ services/frontend/src/components/Overlay/Menu.tsx
@@ -1,5 +1,5 @@
 import Babact from "babact";
-import './menu.css'
+import './overlay.css'
 import { useAuth } from "../../contexts/useAuth";
 import Card from "../../ui/Card";
 import { Link } from "babact-router-dom";
