+++ services/frontend/src/views/OnlineView.tsx
@@ -1,9 +1,7 @@
 import Babact from "babact";
 import { usePong } from "../contexts/usePong";
-import Overlay from "../templates/Overlay";
-import { useNavigate, useParams } from "babact-router-dom";
+import { useParams } from "babact-router-dom";
 import GameOverlay from "../components/Game/GameOverlay";
-import { PongState } from "pong";
 
 export default function OnlineView() {
